Se cambió la ubicación y el nombre de la carpeta temp, ahora es procesados

+++ procesar.sh
@@ -1,33 +1,33 @@
 #!/bin/bash
+#Script Procesar
 
-RUTA=./dataset/descomprimidos/
+RUTA=./dataset/descomprimidos
+RUTA_P=./dataset/procesados
 
-mkdir "$RUTA"temp
-touch "$RUTA"img_names.txt
-touch "$RUTA"img_names_val.txt
+# Verifica que exista el directorio procesados.
+if [[ ! -d $RUTA_P ]]; then
+	mkdir $RUTA_P
+fi
+# Crea los archivos de texto.
+touch $RUTA_P/img_names.txt
+touch $RUTA_P/img_names_val.txt
 IFS=$'\n'
 for FILE in $(ls $RUTA); do 
+	# Lista en el .txt los archivos .jpg.
 	if [[ "${FILE##*.}" == "jpg" ]] ; then
-	       	echo $FILE >> "$RUTA"img_names.txt
+	       	echo $FILE >> $RUTA_P/img_names.txt
 	fi
+	# Lista los archivos con nombre válido en el .txt y copia el archivo.
 	if [[ $FILE =~ ^[A-Z][a-z]+ ]]; then
-		echo $FILE >> "$RUTA"img_names_val.txt
-		cp $RUTA$FILE "$RUTA"temp
+		echo $FILE >> $RUTA_P/img_names_val.txt
+		cp $RUTA/$FILE $RUTA_P
 	fi
 done
-
-cd "$RUTA"temp
+# Procesa las imágenes.
+cd $RUTA_P
 for FIL in $(ls); do
-	#echo $FIL
-	#convert $FIL -gravity center -crop 512x512+0+0 \-extent 512x512 ${FIL%.*}_p.jpg
-	convert $FIL -gravity center -crop 512x512+0+0 \-extent 512x512 p_"${FIL%.*}".jpg
-	rm $FIL
+	if [[ "$FIL" == *.jpg ]]; then
+		convert $FIL -gravity center -crop 512x512+0+0 \-extent 512x512 p_"${FIL%.*}".jpg
+		rm $FIL
+	fi
 done
-sleep 1
-echo $RUTA
-cat "$RUTA"img_names_val.txt
-shopt -s nullglob
-mv "$RUTA"img_names_val.txt "$RUTA"temp
-shopt -u nullglob
-
-
